+++ src/geny_executor/stages/s19_summarize/__init__.py
@@ -9,6 +9,14 @@ from geny_executor.stages.s19_summarize.artifact.default.summarizers import (
     NoSummarizer,
     RuleBasedSummarizer,
 )
+from geny_executor.stages.s19_summarize.frequency_policy import (
+    EveryNTurnsPolicy,
+    FrequencyAwareSummarizerProxy,
+    FrequencyContext,
+    FrequencyPolicy,
+    NeverPolicy,
+    OnContextFillPolicy,
+)
 from geny_executor.stages.s19_summarize.interface import (
     SUMMARY_HISTORY_KEY,
     TURN_SUMMARY_KEY,
@@ -18,10 +26,16 @@ from geny_executor.stages.s19_summarize.interface import (
 from geny_executor.stages.s19_summarize.types import SummaryRecord
 
 __all__ = [
+    "EveryNTurnsPolicy",
     "FixedImportance",
+    "FrequencyAwareSummarizerProxy",
+    "FrequencyContext",
+    "FrequencyPolicy",
     "HeuristicImportance",
     "ImportanceScorer",
+    "NeverPolicy",
     "NoSummarizer",
+    "OnContextFillPolicy",
     "RuleBasedSummarizer",
     "SUMMARY_HISTORY_KEY",
     "SummarizeStage",